Add files via upload
This program was written to help me with my ongoing job hunt for a Data Analyst role. It is capable of scraping available jobs in the position and location of your interest, on Indeed. 

If you're on the lookout for a job, this program will help you save some time by collating all the available jobs.. Try it out!! Feel free to clone and make changes as per your requirement, like a different website etc.